Move application configuration out into its own Configuration class. Use this Configuration class in unit test to remove excess configuration and to actually test the running code.
