+++ src/test/kotlin/au/kilemon/messagequeue/queue/inmemory/InMemoryMultiQueueTest.kt
@@ -1,11 +1,17 @@
 package au.kilemon.messagequeue.queue.inmemory
 
+import au.kilemon.messagequeue.configuration.QueueConfiguration
+import au.kilemon.messagequeue.logging.LoggingConfiguration
 import au.kilemon.messagequeue.queue.AbstractMultiQueueTest
 import au.kilemon.messagequeue.queue.MultiQueue
+import au.kilemon.messagequeue.settings.MessageQueueSettings
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.extension.ExtendWith
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Import
+import org.springframework.context.annotation.Lazy
 import org.springframework.test.context.junit.jupiter.SpringExtension
 
 
@@ -15,26 +21,9 @@ import org.springframework.test.context.junit.jupiter.SpringExtension
  * @author github.com/KyleGonzalez
  */
 @ExtendWith(SpringExtension::class)
-class InMemoryMultiQueueTest: AbstractMultiQueueTest<InMemoryMultiQueue>()
+@Import( *[QueueConfiguration::class, LoggingConfiguration::class, AbstractMultiQueueTest.AbstractMultiQueueTestConfiguration::class] )
+class InMemoryMultiQueueTest: AbstractMultiQueueTest()
 {
-    /**
-     * A Spring configuration that is used for this test class.
-     *
-     * This is specifically creating the [InMemoryMultiQueue] to be autowired in the parent
-     * class and used in all the tests.
-     *
-     * @author github.com/KyleGonzalez
-     */
-    @TestConfiguration
-    internal class InMemoryTestConfiguration
-    {
-        @Bean
-        open fun getInMemoryMultiQueue(): InMemoryMultiQueue
-        {
-            return InMemoryMultiQueue()
-        }
-    }
-
     /**
      * Ensure the [MultiQueue] is cleared before each test.
      */
