+++ src/test/kotlin/au/kilemon/messagequeue/queue/sql/MySqlMultiQueueTest.kt
@@ -1,5 +1,8 @@
 package au.kilemon.messagequeue.queue.sql
 
+import au.kilemon.messagequeue.configuration.QueueConfiguration
+import au.kilemon.messagequeue.logging.LoggingConfiguration
+import au.kilemon.messagequeue.queue.AbstractMultiQueueTest
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
@@ -18,7 +21,7 @@ import java.util.HashMap
  * @author github.com/KyleGonzalez
  */
 @ContextConfiguration(initializers = [MySqlMultiQueueTest.Initializer::class])
-@Import(AbstractSqlMultiQueueTest.SqlMultiQueueTestConfiguration::class)
+@Import( *[QueueConfiguration::class, LoggingConfiguration::class, AbstractMultiQueueTest.AbstractMultiQueueTestConfiguration::class] )
 class MySqlMultiQueueTest : AbstractSqlMultiQueueTest()
 {
     companion object
